refactor: reimport `CurrentUser` param decorator from the new file
Reimport `CurrentUser` param decorator from the new file in
`posts-comments.resolver.ts`, `posts.resolver.ts` and
`users.resolver.ts` files.
